api_yamdb/api/urls.py
from django.urls import include, path

from rest_framework import routers

from api.views import (CategoryViewSet, GenreViewSet, TitleViewSet, UserViewSet, SignupView, TokenView)

router = routers.DefaultRouter()
router.register('categories', CategoryViewSet)
router.register('genres', GenreViewSet)
router.register('titles', TitleViewSet, basename='titles')
router.register(r'users', UserViewSet, basename='users')


urlpatterns = [
    path('v1/', include(router.urls)),
    path('v1/auth/signup/', SignupView.as_view(), name='signup'),
    path('v1/auth/token/', TokenView.as_view(), name='token'),
]
